vue tag exercise
