+++ src/App.vue
@@ -1,36 +1,23 @@
 <script setup>
 import { ref } from 'vue'
 
-let id = 0
-const newTodo = ref('')
-const todos = ref([
-	{id: id++, issue: 'Backlog'}, 
-	{id: id++, issue: 'Todo'},
-	{id: id++, issue: 'Going Hawaii'}
-])
-
-
-
-function addTodo() {
-	todos.value.push({id: id++, issue: newTodo.value})
-}
-
-function removeTodo(todo) {
-	console.log(todo.id)
-	todo.value = todos.value.filter((t) => t !== todo)
-	todos.value.splice(todo.id, 1)
-	// [].filter(t/f)
-}
+const show = ref(true)
+const list = ref([1,2,3])
 
 </script>
+
 <template>
-	<form @submit.prevent="addTodo">
-		<input v-model="newTodo">
-		<button>할 일 추가</button>
-	</form>
-	<ul>
-		<li v-for="todo in todos" v-bind:key="todo">
-			{{ todo.issue }} - <button @click="removeTodo(todo)">Done</button>
-		</li>
-	</ul>
+ <button @click="">List 렌더링 ON/OFF</button>
+ <button @click="list.value.push">List 추가</button>
+ <button @click="remove">List 제거</button>
+ <button @click="reverse">List 뒤집기</button>
+
+ <div class="container">
+	<img id="img" src="./assets/apple.jpg"/>
+ </div>
+ <ul v-if="list.length != 0">
+  <li v-for="item of list">{{ item }}</li>
+ </ul>
+ <p v-else-if="list.length">List is not empty, but hidden.</p>
+ <p v-else>List is empty.</p>
 </template>
