+++ src/assets/apple.jpg:Zone.Identifier
@@ -1,0 +1,4 @@
+[ZoneTransfer]
+ZoneId=3
+ReferrerUrl=https://convertio.co/
+HostUrl=https://s184.convertio.me/p/Wl1Ykc-acml72Af_PNhW8g/92b020e0a40d03fe529e48fff147118a/%EC%82%AC%EA%B3%BC.jpg
